my_test_project/requirements.txt
selenium==4.11.3
requests==2.31.0
pytest==7.2.2
allure-pytest==2.11.10
webdriver-manager==3.8.5